swap improvements
